requirements.txt
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.5.2
async-timeout==5.0.1
asyncio==4.0.0
attrs==25.4.0
blinker==1.9.0
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1648854242877/work
cachetools==6.2.4
ccxt==4.5.33
certifi==2022.6.15.1
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1656782892388/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click==8.1.8
coincurve==21.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
conda==4.14.0
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1649385125392/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1657174295336/work
distro==1.9.0
Django==4.1.2
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-js-asset==2.0.0
django-tinymce==3.5.0
djangorestframework==3.13.1
exceptiongroup==1.3.0
frozenlist==1.8.0
geographiclib==1.52
geopy==2.2.0
gitdb==4.0.12
GitPython==3.1.46
gunicorn==20.1.0
h11==0.16.0
h3==3.7.4
httpcore==1.0.9
httpx==0.28.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llama_stack==0.0.1a5
MarkupSafe==3.0.3
multidict==6.7.0
narwhals==2.15.0
nltk==3.9.2
numpy==1.23.3
packaging==25.0
pandas==2.3.3
Pillow==9.0.1
plotly==6.5.2
propcache==0.4.1
protobuf==6.33.4
psycopg2-binary==2.9.4
pyarrow==21.0.0
pycares==4.11.0
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1649384941891/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2022.5
referencing==0.36.2
regex==2026.1.15
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1661872987712/work
rpds-py==0.27.1
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1653464548430/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.4.3
streamlit==1.50.0
ta==0.11.0
tenacity==9.1.2
termcolor==3.1.0
textblob==0.19.0
timezonefinder==6.1.3
toml==0.10.2
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
tornado==6.5.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
yarl==1.22.0
